Added rest interface and DTO for booking cargo command

+++ booking-module/src/main/java/com/nodamu/cargotracker/booking/domain/valueobjects/BookingAmount.java
@@ -15,5 +15,5 @@ import java.math.BigDecimal;
 @Getter
 @Setter
 public class BookingAmount {
-    private BigDecimal bookingAmount;
+    private Integer bookingAmount;
 }
